# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/auth/signup": {
      "post": {
        "operationId": "AuthController_signup_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignupResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or account already exists."
          }
        },
        "summary": "Register a new account",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/verify-email": {
      "post": {
        "operationId": "AuthController_verifyEmail_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email successfully verified.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyEmailResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired verification link."
          }
        },
        "summary": "Confirm email verification link token",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/stripe-verification": {
      "post": {
        "operationId": "AuthController_completeStripeVerification_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteStripeVerificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stripe verification status updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteStripeVerificationResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed for provided payload."
          },
          "404": {
            "description": "Account not found for provided email."
          }
        },
        "summary": "Record Stripe verification status and dispatch email verification if eligible",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/resend-email": {
      "post": {
        "operationId": "AuthController_resendEmail_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendVerificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification email re-issued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResendVerificationResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Cooldown not elapsed or email invalid."
          }
        },
        "summary": "Resend verification email",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset email dispatched if account exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgotPasswordResultDto"
                }
              }
            }
          }
        },
        "summary": "Request a password reset link",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset completed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetPasswordResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid, expired, or mismatched reset token."
          },
          "404": {
            "description": "Account not found."
          }
        },
        "summary": "Reset account password using a reset link token",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/change-password": {
      "post": {
        "operationId": "AuthController_changePassword_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangePasswordResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or current password is incorrect."
          },
          "401": {
            "description": "Authentication required or token invalid."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change the authenticated account password",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/profile/avatar": {
      "patch": {
        "operationId": "AuthController_updateAvatar_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAvatarDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Avatar updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateAvatarResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid avatar URL supplied."
          },
          "401": {
            "description": "Authentication required."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update the authenticated account avatar",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/profile/notifications": {
      "patch": {
        "operationId": "AuthController_updateNotificationPreferences_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationPreferencesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Notification preferences updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateNotificationPreferencesResultDto"
                }
              }
            }
          },
          "400": {
            "description": "No changes supplied or payload invalid."
          },
          "401": {
            "description": "Authentication required."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update email notification preferences",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login result.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/LoginSuccessResultDto"
                    },
                    {
                      "$ref": "#/components/schemas/LoginVerificationRequiredResultDto"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Credentials are invalid or account requires verification."
          }
        },
        "summary": "Authenticate an account",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/v1/admin/auth/login": {
      "post": {
        "operationId": "AdminAuthController_login_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Admin authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminLoginResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed."
          },
          "401": {
            "description": "Invalid credentials."
          }
        },
        "summary": "Authenticate an admin user",
        "tags": [
          "admin-auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Koajo API",
    "description": "API documentation for Koajo services",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Paste your Koajo access token here"
      }
    },
    "schemas": {
...
      "SignupDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email address for the new account.",
            "example": "user@example.com",
            "maxLength": 320
          },
          "phoneNumber": {
            "type": "string",
            "description": "Contact phone number in E.164 format.",
            "example": "+2348012345678"
          },
          "password": {
            "type": "string",
            "description": "Desired password meeting strong password requirements.",
            "example": "Str0ngP@ssword!",
            "minLength": 8
          },
          "avatarUrl": {
            "type": "object",
            "description": "Optional avatar image URL.",
            "example": "https://cdn.example.com/avatars/user.png",
            "nullable": true
          }
        },
        "required": [
          "email",
          "phoneNumber",
          "password"
        ]
      },
...
      "SignupResultDto": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "description": "Unique identifier for the registered account.",
            "example": "9a4c1d15-3bfe-4f3f-8428-9c2ac5d43052"
          },
          "email": {
            "type": "string",
            "description": "Registered account email address.",
            "example": "user@example.com"
          },
          "phoneNumber": {
            "type": "string",
            "description": "Registered phone number in E.164 format.",
            "example": "+2348012345678"
          },
          "avatarUrl": {
            "type": "object",
            "description": "Avatar image URL if provided during signup.",
            "example": "https://cdn.example.com/avatars/user.png",
            "nullable": true
          },
          "emailVerified": {
            "type": "boolean",
            "description": "Indicates whether the email is verified after signup.",
            "example": false
          },
          "verification": {
            "description": "Verification information for completing signup.",
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/VerificationExpiryDto"
              }
            ]
          }
        },
        "required": [
          "accountId",
          "email",
          "phoneNumber",
          "emailVerified"
        ]
      },
      "VerifyEmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Registered account email address.",
            "example": "user@example.com"
          },
          "token": {
            "type": "string",
            "description": "Verification token included in the email link.",
            "example": "d4c6e8f7ab2c4d1e3f4a5b6c7d8e9f00112233445566778899aabbccddeeff00"
          }
        },
        "required": [
          "email",
          "token"
        ]
      },
      "VerifyEmailResultDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Account email address.",
            "example": "user@example.com"
          },
          "verified": {
            "type": "boolean",
            "description": "Indicates whether the email is now verified.",
            "example": true
          }
        },
        "required": [
          "email",
          "verified"
        ]
      },
      "CompleteStripeVerificationDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Registered account email address.",
            "example": "user@example.com"
          },
          "firstName": {
            "type": "string",
            "description": "First name returned from Stripe identity verification.",
            "example": "Jane",
            "maxLength": 60
          },
          "lastName": {
            "type": "string",
            "description": "Last name returned from Stripe identity verification.",
            "example": "Doe",
            "maxLength": 60
          },
          "stripeVerificationCompleted": {
            "type": "boolean",
            "description": "Indicates whether Stripe identity verification completed successfully.",
            "example": true
          },
          "verificationAttemptCount": {
            "type": "number",
            "description": "Count of Stripe verification attempts recorded.",
            "example": 2
          },
          "verificationFirstAttemptDate": {
            "type": "string",
            "description": "ISO timestamp of the first recorded Stripe verification attempt.",
            "example": "2024-05-01T09:30:00.000Z"
          },
          "verificationLastAttemptDate": {
            "type": "string",
            "description": "ISO timestamp of the most recent Stripe verification attempt.",
            "example": "2024-05-02T16:45:00.000Z"
          },
          "verificationStatus": {
            "type": "string",
            "description": "Current status string reported by Stripe (e.g. `pending`, `verified`).",
            "example": "pending",
            "maxLength": 64
          }
        },
        "required": [
          "email",
          "firstName",
          "lastName",
          "stripeVerificationCompleted"
        ]
      },
...
      "CompleteStripeVerificationResultDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Account email address.",
            "example": "user@example.com"
          },
          "stripeVerificationCompleted": {
            "type": "boolean",
            "description": "Indicates whether Stripe verification has been completed.",
            "example": true
          },
          "verificationAttemptCount": {
            "type": "object",
            "description": "Total number of recorded Stripe verification attempts.",
            "example": 2,
            "nullable": true
          },
          "verificationFirstAttemptDate": {
            "type": "object",
            "description": "ISO timestamp of the first Stripe verification attempt, if known.",
            "example": "2024-05-01T09:30:00.000Z",
            "nullable": true
          },
          "verificationLastAttemptDate": {
            "type": "object",
            "description": "ISO timestamp of the latest Stripe verification attempt, if known.",
            "example": "2024-05-02T16:45:00.000Z",
            "nullable": true
          },
          "verificationStatus": {
            "type": "object",
            "description": "Latest verification status reported by Stripe (e.g. `pending`).",
            "example": "pending",
            "nullable": true
          },
          "verification": {
            "description": "Details about the verification email sent after Stripe completion, if any.",
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/VerificationWindowDto"
              }
            ]
          }
        },
        "required": [
          "email",
          "stripeVerificationCompleted",
          "verificationAttemptCount",
          "verificationFirstAttemptDate",
          "verificationLastAttemptDate",
          "verificationStatus"
        ]
      },
      "ResendVerificationDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Registered account email address.",
            "example": "user@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResendVerificationResultDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Account email address.",
            "example": "user@example.com"
          },
          "verification": {
            "description": "Details about the freshly issued verification email.",
            "allOf": [
              {
                "$ref": "#/components/schemas/VerificationWindowDto"
              }
            ]
          }
        },
        "required": [
          "email",
          "verification"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email associated with the account.",
            "example": "user@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "ForgotPasswordResultDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Account email address.",
            "example": "user@example.com"
          },
          "requested": {
            "type": "boolean",
            "description": "Indicates that the reset request was registered.",
            "example": true
          }
        },
        "required": [
          "email",
          "requested"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email associated with the account.",
            "example": "user@example.com"
          },
          "token": {
            "type": "string",
            "description": "Password reset token supplied in the email link.",
            "example": "d4c6e8f7ab2c4d1e3f4a5b6c7d8e9f00112233445566778899aabbccddeeff00"
          },
          "newPassword": {
            "type": "string",
            "description": "New password that meets strong password requirements.",
            "example": "N3wStr0ngP@ssword!",
            "minLength": 8
          }
        },
        "required": [
          "email",
          "token",
          "newPassword"
        ]
      },
      "ResetPasswordResultDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Account email address.",
            "example": "user@example.com"
          },
          "reset": {
            "type": "boolean",
            "description": "Indicates that the password was successfully reset.",
            "example": true
          }
        },
        "required": [
          "email",
          "reset"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "description": "Current password for the authenticated account.",
            "example": "Curr3ntP@ss!",
            "minLength": 8
          },
          "newPassword": {
            "type": "string",
            "description": "Desired new password meeting strong password requirements.",
            "example": "N3wStr0ngP@ssword!",
            "minLength": 8
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "ChangePasswordResultDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indicates that the password update was successful.",
            "example": true
          }
        },
        "required": [
          "success"
        ]
      },
      "UpdateAvatarDto": {
        "type": "object",
        "properties": {
          "avatarUrl": {
            "type": "object",
            "description": "Avatar image URL. Send null to clear the avatar.",
            "example": "https://cdn.example.com/avatars/user.png",
            "nullable": true
          }
        }
      },
      "UpdateAvatarResultDto": {
        "type": "object",
        "properties": {
          "avatarUrl": {
            "type": "object",
            "description": "Updated avatar URL. Null means the avatar has been cleared.",
            "example": "https://cdn.example.com/avatars/user.png",
            "nullable": true
          }
        },
        "required": [
          "avatarUrl"
        ]
      },
      "UpdateNotificationPreferencesDto": {
        "type": "object",
        "properties": {
          "emailNotificationsEnabled": {
            "type": "boolean",
            "description": "Enable or disable general system emails.",
            "example": true
          },
          "transactionNotificationsEnabled": {
            "type": "boolean",
            "description": "Enable or disable transaction-related emails.",
            "example": true
          }
        }
      },
      "UpdateNotificationPreferencesResultDto": {
        "type": "object",
        "properties": {
          "emailNotificationsEnabled": {
            "type": "boolean",
            "description": "Indicates if general system emails are enabled.",
            "example": true
          },
          "transactionNotificationsEnabled": {
            "type": "boolean",
            "description": "Indicates if transaction-related emails are enabled.",
            "example": true
          }
        },
        "required": [
          "emailNotificationsEnabled",
          "transactionNotificationsEnabled"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Account email address.",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "Account password used for authentication.",
            "example": "Str0ngP@ssword!",
            "minLength": 8
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
    }
  }
}